Fixed single peer to peer example
