+++ Network/Source/Scenes/SinglePeerToPeerExample.ts
@@ -1,0 +1,98 @@
+import * as FudgeNetwork from "../ModuleCollector";
+
+let asMode: boolean = false;
+const networkClient: FudgeNetwork.NetworkClientManager = new FudgeNetwork.NetworkClientManager();
+const peerToPeerSignalingServer: FudgeNetwork.SinglePeerSignalingServer = new FudgeNetwork.SinglePeerSignalingServer();
+
+
+
+FudgeNetwork.UiElementHandler.getAllUiElements();
+FudgeNetwork.UiElementHandler.startSignalingButton.addEventListener("click", startingUpSignalingServer);
+FudgeNetwork.UiElementHandler.signalingSubmit.addEventListener("click", connectToSignalingServer);
+FudgeNetwork.UiElementHandler.loginButton.addEventListener("click", createLoginRequestWithUsername);
+FudgeNetwork.UiElementHandler.connectToUserButton.addEventListener("click", connectToOtherPeer);
+FudgeNetwork.UiElementHandler.sendMsgButton.addEventListener("click", sendMessageViaPeerConnectionChannel);
+FudgeNetwork.UiElementHandler.stopSignalingServer.addEventListener("click", turnOffSignalingServer);
+
+
+function addKeypressListener(): void {
+    let browser: Document = FudgeNetwork.UiElementHandler.electronWindow;
+    browser.addEventListener("keydown", (event: KeyboardEvent) => {
+        if (event.keyCode == 27) {
+            networkClient.sendDisconnectRequest();
+        }
+        else {
+            networkClient.sendKeyPress(event.keyCode);
+        }
+    });
+}
+
+function createLoginRequestWithUsername(): void {
+    let chosenUserName: string = "";
+    if (FudgeNetwork.UiElementHandler.loginNameInput) {
+        chosenUserName = FudgeNetwork.UiElementHandler.loginNameInput.value;
+        console.log("Username:" + chosenUserName);
+        networkClient.checkChosenUsernameAndCreateLoginRequest(chosenUserName);
+    }
+    else {
+        console.error("UI element missing: Loginname Input field");
+    }
+}
+
+function connectToOtherPeer(): void {
+    let userNameToConnectTo: string = "";
+    if (FudgeNetwork.UiElementHandler.usernameToConnectTo) {
+        userNameToConnectTo = FudgeNetwork.UiElementHandler.usernameToConnectTo.value;
+        networkClient.checkUsernameToConnectToAndInitiateConnection(userNameToConnectTo);
+    }
+    else {
+        console.error("Missing Ui Element: Username to connect to");
+    }
+}
+
+function sendMessageViaPeerConnectionChannel(): void {
+    let messageToSend: string = FudgeNetwork.UiElementHandler.msgInput.value;
+    FudgeNetwork.UiElementHandler.chatbox.innerHTML += "\n" + networkClient.localUserName + ": " + messageToSend;
+    networkClient.sendMessageToSingularPeer(messageToSend);
+}
+
+
+
+function startingUpSignalingServer(): void {
+    console.log("Turning server ONLINE");
+
+    peerToPeerSignalingServer.startUpServer(8080);
+
+    let startSignalingButton: HTMLButtonElement = FudgeNetwork.UiElementHandler.startSignalingButton as HTMLButtonElement;
+    startSignalingButton.hidden = true;
+    let stopSignalingButton: HTMLButtonElement = FudgeNetwork.UiElementHandler.stopSignalingServer as HTMLButtonElement;
+    stopSignalingButton.hidden = false;
+}
+
+function turnOffSignalingServer(): void {
+    peerToPeerSignalingServer.closeDownServer();
+
+
+    let startSignalingButton: HTMLButtonElement = FudgeNetwork.UiElementHandler.startSignalingButton as HTMLButtonElement;
+    startSignalingButton.hidden = false;
+    let stopSignalingButton: HTMLButtonElement = FudgeNetwork.UiElementHandler.stopSignalingServer as HTMLButtonElement;
+    stopSignalingButton.hidden = true;
+    let switchButton: HTMLButtonElement = FudgeNetwork.UiElementHandler.switchModeButton as HTMLButtonElement;
+    switchButton.hidden = false;
+}
+function connectToSignalingServer(): void {
+    networkClient.signalingServerConnectionUrl = "ws://" + FudgeNetwork.UiElementHandler.signalingUrl.value;
+    networkClient.connectToSignalingServer();
+
+    addKeypressListener();
+}
+
+
+
+// Changing HTML pages restarts the renderer process, causing connection loss on networking
+// so not doable this way. Single page required or different way to change page (testing only so not that important)
+// const { remote } = require('electron')
+// function testbutton () {
+//     remote.getCurrentWindow().loadURL('https://github.com')
+// }
+// FudgeNetwork.UiElementHandler.signalingSubmit.addEventListener("click", testbutton);
